fix hf 14a sim,   on device side a sneaky detection if data started with zero,  it would treat it as uid is in emulator memory. Also fixed a size issue on flag

+++ armsrc/Standalone/hf_aveful.c
@@ -248,7 +248,7 @@ void RunMod(void) {
                     state = STATE_SEARCH;
                 }
             } else if (state == STATE_EMUL) {
-                uint8_t flags = FLAG_7B_UID_IN_DATA;
+                uint16_t flags = FLAG_7B_UID_IN_DATA;
 
                 Dbprintf("Starting simulation, press pm3-button to stop and go back to search state.");
                 SimulateIso14443aTag(7, flags, card.uid, 0);
